projects/polymarket/polyquantbot/phase4/requirements.txt
aiohttp>=3.9.0
aiosqlite>=0.19.0
yaml-config>=0.1.4
pyyaml>=6.0.1
structlog>=24.1.0
httpx>=0.27.0
python-dotenv>=1.0.0
uuid>=1.30
